fix: Rewrite audio streaming tests to match actual SDK API, fix CRC32 compilation
- Move tests from test-jvm/ (unrecognized) to jvmTest/ (correct KMP source set)
- Rewrite all 3 test files against actual AudioChunk/BleAudioStreamer/Extensions API
- Fix CRC32 Int/Long type errors in AudioChunkStreamingProtocol and BleAudioStreamer
- Fix pre-existing SDKTest.kt (referenced removed API methods)
- Remove root tests-jvm/ module and duplicate audio stubs
- Remove repo_cache submodule, restore MARISOL.md, revert .gitignore
- Expand CI to run SDK JVM tests alongside watch tests
- 53 tests passing (14 BleAudioStreamer, 15 Extensions, 23 Protocol, 1 SDK)

+++ sdk/runanywhere-kotlin/src/commonMain/kotlin/com/runanywhere/sdk/features/audio/streaming/BleAudioStreamer.kt
@@ -65,13 +65,13 @@ abstract class BleAudioStreamer(
             crc = crc xor (byte.toInt() and 0xFF)
             for (bit in 0..7) {
                 if ((crc and 1) != 0) {
-                    crc = (crc ushr 1) xor 0xEDB88320
+                    crc = (crc ushr 1) xor 0xEDB88320.toInt()
                 } else {
                     crc = crc ushr 1
                 }
             }
         }
-        return crc xor 0xFFFFFFFF
+        return crc xor 0xFFFFFFFF.toInt()
     }
     
     /**
